Conflicts:
	app/assets/stylesheets/email.css.less
	app/views/activity_mailer/activities/_added_evidence.html.erb
	app/views/activity_mailer/activities/_followed_user.html.erb

+++ core/bin/ci/203_acceptance_tests.sh
@@ -1,4 +1,6 @@
 #!/bin/bash
+
+[ -z $SUPPRESS_TESTING ] || exit 0
 echo "Running acceptance tests"
 
 rm -rf tmp/capybara/
@@ -13,14 +15,11 @@ function do_tests {
   test ${PIPESTATUS[0]} -eq 0 || touch TEST_FAILURE
 }
 
-
 do_tests
 if grep -qe 'PhantomJS has crashed.' < $OUTPUTFILE ; then
   echo "Detected random fail, retrying"
   do_tests
 fi
-
-
 
 if ! grep -qe '<testcase' < $REPORTFILE ; then
   echo "FAILING BUILD: No testcases found in $REPORTFILE"
